+++ CHANGES.md
@@ -39,6 +39,7 @@ Improvements 🙌:
  - Add graphic resources for F-Droid (#812, #2220)
  - Highlight text in the body of the displayed result (#2200)
  - Considerably faster QR-code bitmap generation (#2331)
+ - Open an existing DM instead of creating a new one (#2319)
 
 Bugfix 🐛:
  - Fixed ringtone handling (#2100 & #2246)